Isolate server code
